requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
psycopg==3.2.6
psycopg-binary==3.2.6
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.13.2
